HW12 is completed
