+++ hw12_13_14_15_calendar/internal/app/app.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/andreyAKor/otus_hw/hw12_13_14_15_calendar/internal/repository"
+	"github.com/andreyAKor/otus_hw/hw12_13_14_15_calendar/internal/repository/repository"
 
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog/log"
