+++ hw12_13_14_15_calendar/internal/repository/repository/repository.go
@@ -1,0 +1,45 @@
+package repository
+
+import (
+	"context"
+	"io"
+	"time"
+
+	"github.com/pkg/errors"
+)
+
+var (
+	ErrTimeBusy = errors.New("event at this time is busy")
+	ErrNotFound = errors.New("event not found")
+)
+
+type EventsRepo interface {
+	Create(ctx context.Context, ev Event) (int64, error)
+	Update(ctx context.Context, id int64, ev Event) error
+
+	Delete(ctx context.Context, id int64) error
+	DeleteOld(ctx context.Context) error
+
+	GetListByDate(ctx context.Context, date time.Time) ([]Event, error)
+	GetListByWeek(ctx context.Context, start time.Time) ([]Event, error)
+	GetListByMonth(ctx context.Context, start time.Time) ([]Event, error)
+}
+
+type DBEventsRepo interface {
+	Connect(ctx context.Context, dsn string) error
+
+	io.Closer
+	EventsRepo
+}
+
+type Event struct {
+	ID            int64
+	Title         string
+	Date          time.Time
+	Duration      time.Duration
+	Descr         *string
+	UserID        int64
+	DurationStart *time.Duration
+	CreatedAt     time.Time
+	UpdatedAt     *time.Time
+}
